+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ### Build tool
 
+- The build tool now generates Hexdocs URLs using the new format of
+  `package.hexdocs.pm` rather than `hexdocs.pm/package`.
+  ([Surya Rose](https://github.com/GearsDatapacks))
+
 ### Language server
 
 ### Formatter